requirements.txt
Cython
librosa==0.9.1
matplotlib
numpy
phonemizer
scikit-learn==1.0.2
scipy
tensorboard
torch
torchvision
torchaudio
unidecode
amfm_decompy
tqdm
cn2an
pypinyin
jieba
transformers
demucs
openai-whisper
imageio==2.4.1
moviepy
gradio
fastapi
